xf86-video-freedreno: upgrade to 1.3.0
Since version 1.2.0 this video driver has support for:

==
 * platform-probe and autoconfig support, so no .conf file needed
    with xserver 1.16+
==

However this requires Xorg 1.16 to work. As such we continue to include a custom
xorg.conf file, and this file is likely to break on non-freedreno
platforms. When Xorg is migrated to 1.16, we can remove the custom configuration
file, and solely rely on platform probe.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>