dragonboard410: align rootfs size on 4096
when the .ext4 raw image is not a multiple of 4096 , fastboot will fail to cut
it into chunks when running 'fastboot flash' command.

If the image size is less than the intermediate buffer that fastboot uses to
copy the image file from host, then it won't cut it into chunks, and in that
case the alignment is not really needed, but to be safe, let's make sure all
images for this board are aligned properly.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>