qrtr: ensure prefix variable is set properly
prefix variable is set in makefile, but also in the environment by OE. Until
krogoth make was always called with -e, so the variable was set based on the
environment. However since OE core commit
aeb653861a0ec39ea7a014c0622980edcbf653fa, -e has been removed, so we need to set
prefix in the recipe to override the value from the Makefile.

It is not doing anything harmful on jethro, so let's apply this change on all
branches.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>