gpsd-machine-conf: move bbappend file into layer specific folder
If meta-qcom is used in a configuration without meta-oe, bitbake will fail to
parse with this error:

ERROR: No recipes available for:
  ../build-nodistro/conf/../../layers/meta-qcom/recipes-navigation/gpsd/gpsd-machine-conf_%.bbappend

The layer.conf file has logic to selectively add .bb and .bbappend from layer
specific folders, so let's move this bbappend into the 'openembedded-layer', so
that it will be parsed only if 'openembedded-layer' (e.g. meta-oe) is being used.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>