linux-linaro-qcomlt: remove use of KERNEL_OUTPUT
KERNEL_OUTPUT used to be set in kernel.bbclass as the kernel output file, and it
was used for convenience here. However in OE-core commit
849b67b2e4820564b5e5c9bd4bb293c44351c5f3 the definition of KERNEL_OUTPUT was
removed (for legit reasons), and our recipe no longer works. Since we use
KERNEL_OUTPUT in the do_compile() step it is actually fine (and even simpler) to
directly use the kernel binary anyways.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>