linux-qcom-bootimg: create missing symlink for the boot image
Before commit 589bbe8bf, we would create a symlink for the deployed boot image
as boot-${MACHINE}.img. This commit reworked how the variables were set and we
missed the fact that this symlink is no longer created, and instead we have
boot-${KERNEL_IMAGETYPE}-${MACHINE}.img, so let's add the inital symlink back to
avoid breakage.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>