gstreamer1.0-plugins-bad: fix the logic to set GL option in -bad
The purpose of this .bbappend file is to enable the right GL config option in
Gstreamer -bad plugins for qcom based chipset. The current logic was broken:
when building with no x11 and no wayland in DISTRO_FEATURES we would set
PACKAGECONFIG_GL to opengl which is wrong, since opengl depends on x11 (in
-bad). So the right thing to do is to set it to opengl is x11 is in
DISTRO_FEATURES and gles2 otherwise, e.g. when building for wayland only (no
X11) and for use cases such as Qt5/EGLFS that does not use x11 nor wayland.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>