qmi-gps-proxy: move the recipe in openembedded-layer sublayer
This recipe explicitely depends on gpsd which is in meta-oe and not OE-core. So
let's move the recipe into the openembedded-layer so that it gets parsed/used
only when our users also have enabled meta-oe in their build, otherwise this
recipe is irrelevant and should be skipped.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
(cherry picked from commit 061ab4a573353690c9ab50e59d8bbcb6a54269fa)