firmware-qcom-dragonboard410c: add symlinks for venus files
Venus firmware files have reached linux-firmware git tree, but they are located
at a different location in /lib/firmware. Depending on which kernel version the
users are compiling it might look for them in the 'old' or the 'new'
location. For now add symlinks, until we switch to newer kernel and start using
linux-firmware package.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>