firmware: use firmware from meta-qcom
Some firmware blob for APQ8016 are now included in linux-firmware, however not
all of them. It creates conflicts when trying to install 'linux-firmware' and
'firmware-qcom-dragonboard410c' packages.

This patch 'backports' the linux-firmware packaging for QCOM firmware which was
merged in OE-core:

687d71982a linux-firmware: add packaging for QCOM firmware for APQ8016 and APQ8096

Only the bits for APQ8016 are backported, since oe-core/rocko linux-firmware
recipe is still a bit too old, and does not include the APQ8096 firmware (yet).

With this change some firmware are packaged twice then, and for now we want to
give priority to the firmware recipe from meta-qcom, and we use
RCONFLICTS/RREPLACES/RPROVIDES for that.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>