kernel: append DTB to kernel image
Instead of using dtbTool to create a DT image , and include it into the boot
image, we append DTB to the compressed kernel image. the LK bootloader knows how
to retrieve DTB , when they are appeneded to compressed kernel image.

Note:
* it only works with compressed image.
* dtbTool also typically adds msm-ids into the DTB, recent versions of LK exist
that allow a single DTB to be used without the msm-ids, in which case it will
work fine. when using older LK version, it is needed to patch the kernel and
include the msm-ids.
* QCOM_BOOTIMG_BUNDLE_DT can be unset , since we longer need the DT image to
boot.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
(cherry picked from commit b97de48b1f58e339b2d40098d3369acf9b94b30f)