qrtr: include qrtr and rmtfs by default
qrtr and rmtfs are very standard user space helpers when using Qualcomm
Snapdragon processors. They are used to manage services running on remote cores,
such as modem, DSP, ... While they are needed for GPS, they could be needed for
other use cases, so let's add them by default from now on, insead of as a
dependency of GPSd.

While at it, at them for 8016 and 8096 targets.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>