boot-init: Invoke bootrr scripts automatically
In a LAVA-less environment it's convenient to have the bootrr board
scripts to be automatically invoked during boot. Allow this to happen
by checking the kernel command line for the magic "bootrr-auto" and if
so invoke the relevant board scripts.

Finish with writing the tilde sentinel, to provide a signal to
automation tools of when the tests are done executing.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>