layer.conf: include special treatment for meta-backports
meta-backports is a layer that contains various backport from existing OE
layers, most of the time to bring newer version of recipes into an OE stable
branch.

Some meta-qcom users will also use meta-backports, some won't. This patch adds a
bit of logic when setting BBFILES so that we can hold patches in meta-qcom that
are parsed/used only when meta-backports is also used.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>