recipes-test: Add a new initramfs-test-full-image.bbappend
'initramfs-test-image' is used across several boards and changes
to the same are very keenly looked at for size changes to the overall
image thus produced as it may lead to some boards failing to boot with
yocto based images.

So it makes sense to introduce a new 'initramfs-test-full-image' which
can retain the character of the test image but can house more utilities
which are useful for several qcom boards.

So, we have now three types of initramfs images:
- initramfs-tiny-image,
- initramfs-test-image, and
- initramfs-test-full-image

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@linaro.org>
(cherry picked from commit db16ed568d495266ac326fb05ea75ebbe7886473)
[ replaced dhcpd with dhcp-client in dunfell ]
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>