bootrr, diag: make sure branch and protocol are configured
The default branch name for new repositories created on GitHub is now
main. While it does not affect existing repo, let's be more explicit
just in case, as there is uncertainty about how it will be managed in
general.

Also Github has announced that it will stop supporting git:// protocol
in Q1 2022, let's make sure we don't use it anymore, and use
https:// instead.

https://github.blog/2021-09-01-improving-git-protocol-security-github/#no-more-unauthenticated-git

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>