initramfs-test-image: generify PACKAGE_INSTALL handling
For the initramfs-test-image family of images we have to simulate
dynamic layers/bbappends support in the form of per-layer variables.
Currently only two layers are supported this way: openembedded-layer
(meta-oe) and networking-layer (meta-networking), both are handled
manually. Allow initramfs-*-image to provide dynamic package lists for
any layer, by looping through the BBFILE_COLLECTIONS in an anonymous
Python function. This causes minor change in PACKAGE_INSTALL_foo-layer
variables syntax.

Also as we at it, move this dynamic handling to the
initramfs-tiny-image, so that any image branching from it can use this
functionality.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>