packagegroup-qcom-boot: group modules required to find rootfs
Add a packagegroup which would recommend kernel modules (and, if required,
other packages) necessary to find rootfs on supported Qualcomm boards.

Probably it is not the full list (ext4 is not included for example), but
it lists packages required to boot when using the vanilla kernel
defconfig for arm64 platform.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>