.github: use pull_request_target to allow PR testing
When using pull_request event, the workflow is run in the context of
the PR, so it's potentially untrusted code from a third party. As such
Github Action secrets are not available for security reason. The PR
build test workflow requires a secret (TUXSUITE TOKEN). So instead, we
can use pull_request_target which ensures the workflow runs in the
context of the base of the PR, e.g. trusted/merged code. In this case
the repo secrets can be used in the workflow.

We just need to carefully set the 'ref' we use in the tuxOE job
definition, which needs to refer to the PR.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>