abl: add recipe to build (unsigned) ABL binary
Add a recipe to build an unsigned ABL binary. This can be useful to test
the ABL changes. Currently this recipe uses external Clang binary, but
it might be later migrated to meta-clang. The ABL build system depend on
Python2 being present, thus the recipe will not be enabled if
meta-python2 is not present.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>