.github: add ref_type param
For daily builds, using the 'schedule' trigger, the default branch is
being used only. So we need to be able to trigger builds by 'branch'
instead of by 'ref'. ref_type is a new input parameter, so that
calling jobs can decide to use a ref (SHA) or a branch.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>