android-gadget-setup: fix conditional argument to assign serial variable
Change the conditional argument to check if androidserial is non-empty. The
condition `[ -n ]` when `androidserial` is empty is equated to `false` which
results in an non-zero status from `android-gadget-setup.machine`.
/usr/bin/android-gadget-setup starts with `set -e` is used to change
behavior of the shell in various ways. `-e` option causes the shell to exit as
soon a command exits with a non-zero status. Hence, executing `true` in
the end results in a zero exit status in all cases, which suffices the
condition  for `set -e`, hence unblocking android-tools-adbd service when
`androidserial` is not defined.

Signed-off-by: Chirag Jain <quic_chirjain@quicinc.com>