firmware-woa: fix cabinets download
The cabinet files do not have platform and/or version embedded into the
filename. This way there is a huge possiblity of the clash if two
platforms use the same SoC. Also updating the PV doesn't cause cabinets
being redownloaded. Fix this by encoding path and PV into the downloaded
file name.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>