esp-qcom-image: remove unused dirs and files
OE creates a set of directories and files which are desired on standard
images, but are not required for ESP as this is a minimal non-rootfs
standard image.

Remove everything but the 'EFI' folder from the ESP image via a custom
IMAGE_PREPROCESS_COMMAND command, as the base files and directories
are needed during the rootfs task.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>